requirements.txt
psycopg2==2.5.4
django==1.7
django-allauth==0.18.0
django-formsettesthelpers==0.1
djangorestframework==3.0.2
markdown==2.5.1
django-filter==0.8
